USBAI-C++/requirements.txt
# Core dependencies
faiss-cpu==1.12.0
sentence-transformers==2.7.0
torch==2.6.0
transformers==4.35.2
llama-cpp-python==0.2.11
pydub==0.25.1
pytesseract==0.3.10
python-docx==1.1.0
# PyMuPDF==1.23.8  # Requires Visual Studio Build Tools - install manually if needed
open_clip_torch==2.24.0
numpy==1.24.3
flask==2.3.3
python-multipart==0.0.6
pillow==10.0.1
librosa==0.10.1
soundfile==0.12.1
sqlite3==3.42.0
